// src/app/Models/product.ts
export class Product {
  id: number = 0;
  name: string = '';
  category: string = '';
  image: string = '';
  price: number = 0;
  oldPrice: number = 0;
  discount: number = 0;
  rating: number = 0;
  reviews: number = 0;
  inStock: boolean = false;
  description: string = '';
  highlights: string[] = [];
}